+++ src/components/Section.jsx
@@ -1,0 +1,10 @@
+
+
+export default function Section({children, title, id}) {
+    return(
+        <section id={id}>
+            <h2>{title}</h2>
+            {children}
+        </section>
+    )
+}
